Revert "Add API code"
